+++ extensions/pipe-commands/assets/commands/change-case/uppercase.py
@@ -4,7 +4,7 @@
 # @raycast.title Switch to Uppercase
 # @raycast.packageName Text Actions
 # @raycast.mode pipe
-# @raycast.input { "type": "text" }
+# @raycast.inputType text
 
 import sys