+++ extensions/pipe-commands/assets/commands/file/copy-path.sh
@@ -4,6 +4,6 @@
 # @raycast.title Copy Path
 # @raycast.packageName File Actions
 # @raycast.mode pipe
-# @raycast.input { "type": "file" }
+# @raycast.inputType file
 
 cat