+++ extensions/pipe-commands/assets/commands/text/convert-space-to-newline.sh
@@ -4,7 +4,7 @@
 # @raycast.title Convert Space to Newline
 # @raycast.packageName Text Actions
 # @raycast.mode pipe
-# @raycast.input { "type": "text" }
+# @raycast.inputType text
 # @raycast.icon 🔤
 
 tr ' ' '\n'