+++ extensions/pipe-commands/assets/commands/text/reverse-lines.sh
@@ -4,7 +4,7 @@
 # @raycast.title Reverse Lines
 # @raycast.packageName Text Actions
 # @raycast.mode pipe
-# @raycast.input { "type": "text" }
+# @raycast.inputType text
 # @raycast.icon 🔤
 
 tail -r